+++ vertigo-core/src/test/java/io/vertigo/core/analytics/process/ProcessAnalyticsTest.java
@@ -1,7 +1,7 @@
 /**
  * vertigo - application development platform
  *
- * Copyright (C) 2013-2022, Vertigo.io, team@vertigo.io
+ * Copyright (C) 2013-2023, Vertigo.io, team@vertigo.io
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -204,5 +204,25 @@ public final class ProcessAnalyticsTest extends AbstractTestCaseJU5 {
 					}
 				});
 		log.trace("elapsed = " + (System.currentTimeMillis() - start));
+	}
+
+	@Test
+	public void testAddTag() {
+		analyticsManager.trace(
+				TEST_CATEGORY,
+				"/create/orders",
+				tracer -> {
+					tracer.addTag("testTag", "MyTag");
+				});
+	}
+
+	@Test
+	public void testAddMetadata() {
+		analyticsManager.trace(
+				TEST_CATEGORY,
+				"/create/orders",
+				tracer -> {
+					tracer.addMetadata("testMetaData", "MyMetaData");
+				});
 	}
 }